Utilities: Move DateCollider into being a utility and not just an object.
